// src/app/models/User.model.ts
export interface User {

  // constructor(userName: string, password: string, email: string, dateOfBirth: Date, residency: string) {
  //   this.UserName = userName;
  //   this.Password = password;
  //   this.Email = email;
  //   this.DateOfBirth = dateOfBirth;
  //   this.Residency = residency;
  // }
  UserName: string;
  Password: string;
  Email: string;
  DateOfBirth: Date;
  Residency: string;
}